perbaikan

+++ database/migrations/2024_03_12_140349_create_trans_development_children_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('student_id')->index();
             $table->foreignId('development_childerns_id');
-            $table->double('score',2,2);
+            $table->integer('score');
             $table->string('assessment_from')->nullable()->index();
             $table->string('validasi')->default('0')->index();
             $table->timestamps();
